Rename: `actionsFilter` -> `actionsSort`
